fix: refactor getIpAddress
