--- services/utils/getIpAddress.php
@@ -1,13 +1,0 @@
-<?php
-    function getIpAddress()
-    {
-        if (!empty($_SERVER['HTTP_CLIENT_IP'])) { //check ip from share internet
-            $ipaddress = $_SERVER['HTTP_CLIENT_IP'];
-        } elseif (!empty($_SERVER['HTTP_X_FORWARDED_FOR'])) { // proxy pass ip
-            $ipaddress = $_SERVER['HTTP_X_FORWARDED_FOR'];
-        } else {
-            $ipaddress = $_SERVER['REMOTE_ADDR'];
-        }
-        list($ipaddress) = explode(',', $ipaddress);
-        return $ipaddress;
-    }
